Result of `b3DebugDraw.DrawShapeFcn` is ignored Fixes #76 Several issues with the sample `CharacterMover` Fixes #75
