+++ samples/CMakeLists.txt
@@ -155,6 +155,8 @@ set(SAMPLE_FILES
     sample.h
     mesh_loader.cpp
     mesh_loader.h
+    mover.cpp
+    mover.h
     jsmn.h
     tiny_obj_loader.h
     earcut.h
