+++ src/types.c
@@ -80,10 +80,9 @@ b3ShapeDef b3DefaultShapeDef( void )
 	return def;
 }
 
-static bool b3EmptyDrawShape( void* userShape, b3WorldTransform transform, b3HexColor color, void* context )
+static void b3EmptyDrawShape( void* userShape, b3WorldTransform transform, b3HexColor color, void* context )
 {
 	B3_UNUSED( userShape, transform, color, context );
-	return false;
 }
 
 static void b3EmptyDrawSegment( b3Pos p1, b3Pos p2, b3HexColor color, void* context )
